Applied base-0031-ext-timeout
When an ext_ring (slave FSM) datagram does not fit in the current cycle, it
is checked for timeout based on how long it has been since the datagram was
populated -- except that this origin time was never actually set, resulting
in checking the time since that datagram *slot* was last used to send some
other datagram.  This results in all such datagrams being timed out instead
of being deferred to the next cycle as expected.

(Triggered by having a large domain and lots of outstanding slave requests,
combined with a fast enough send_interval to not be able to queue all of
them; this is also affected by the slave mailbox size.)