Fix ethercat tool compilation
Fixes compilation error:
  /usr/bin/ld: error: ethercat uses VFP register arguments, ../master/soe_errors.o does not
  /usr/bin/ld: failed to merge target specific data of file ../master/soe_errors.o
  collect2: ld returned 1 exit status

The error occurs due to different compiler options when building
kernel modules and tools. When building the ethercat utility,
the file soe_errors.o is linked directly from the kernel module
build directory. To fix this problem, we just copy the source
file soe_errors.c into the build directory of the ethercat utility
and build it together with the utility.

This is certainly not quite right, but it is a quick fix.

Signed-off-by: Anton Kikin <a.kikin@tano-systems.com>