[PATCH 2/2] 0002-retry-dc-register
Hot plugged slaves can fail to read DC registers if they are read before
the EEPROM is loaded.

In the EtherCAT Hardware Data Sheet Section I document, section 17.2
Power-on Sequence, it says in Step 4:
'Process memory is not accessible until the SII EEPROM is loaded,
as well as any function depending on ESC Configuration data'

For various ESC's (e.g. ET1201 ASIC) activation of the DC unit is based
on bits .2 and .3 of register 0x0141, which is loaded from the EEPROM.
While the DC unit is not activated the DC registers can not be read or
written from the master.

Section 17.2, Step 6, says that once the EEPROM is loaded, Register 0x0110
bit 0 turns to 1.  Due to this I have reordered the fsm_slave_scan
states so that the datalink states (which read register 0x0110) are run
before checking the dc registers.  This state now blocks until bit 0 is
set and error's out if there is a timout.

Apparently ET1100 and ET1201 ESC's get their DC units activated based on
information read from the EEPROM.  Only the ET1200 has the DC unit
activated by default, allowing the DC registers to be read immediately.

This patch reorders the slave scan state machine from:
  Base --> [ DCCap ] --> DataLink --> SII
To:
  Base --> DataLink --> [ DCCap ] --> SII

Where the DataLink state will block until the EEPROM read is complete (with a 500ms timeout).