fix: if ${MODPROBE_FLAGS} is empty, it confuses modprobe when in ""
When running shellcheck, it complained about possible globbing in
${MODPRPOBE_FLAGS}, which is either undefined or externally set. In the
case it is empty / undefined, it evaluates to '', hence modprobe is
trying to load the module with empty name, which is not found btw (spoiler).

Instead of adding code that checks if MODPROBE_FLAGS is empty / undefined,
I just removed it since I guess it is not used, as it is not documented
either.

Signed-off-by: Matthias Schoepfer <m.schoepfer@rethinkrobotics.com>