Fix "No such file or directory" error when using multiple masters and RTDM
This error occurs when we are using multiple masters along with RTDM.
Device minor is not set and device numbers are always one when registering
an RTDM device.
Seems there is no need to change the legacy Xenomai 2.6 RTDM driver, because
the device will be identified by a name instead of a device minor.