lib: slave_config: Fix inline issue for ec_slave_config_add_soe_request
When calling ec_slave_config_add_soe_request in order place,
some GCC versions perform inline substitution as the part of optimisation,
so this can lead to the absence of an actual function call in the compiled code.
This patch fixes the inline building issue for ec_slave_config_add_soe_request.

Signed-off-by: Long Luu <long.luu.ur@renesas.com>